Move performance tests out of unit_tests
